+++ aiorequestful/auth/__init__.py
@@ -1,1 +1,2 @@
-from ._base import Authoriser, AuthRequest, AuthResponseHandler, AuthResponseTester
+from ._base import Authoriser
+from ._utils import AuthRequest, AuthResponseHandler, AuthResponseTester, SocketHandler
